+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/Camera/HDCamera.cs
@@ -64,7 +64,7 @@ namespace UnityEngine.Rendering.HighDefinition
         public bool colorPyramidHistoryIsValid = false;
         public bool volumetricHistoryIsValid   = false; // Contains garbage otherwise
         public int  colorPyramidHistoryMipCount = 0;
-        public VBufferParameters[] vBufferParams; // Double-buffered
+        public VBufferParameters[] vBufferParams; // Double-buffered; needed even if reprojection is off
 
         float m_AmbientOcclusionResolutionScale = 0.0f; // Factor used to track if history should be reallocated for Ambient Occlusion
 
@@ -263,11 +263,14 @@ namespace UnityEngine.Rendering.HighDefinition
             return m_NeedTAAResetHistory;
         }
 
-        public bool IsVolumetricReprojectionEnabled()
+        internal bool IsVolumetricReprojectionEnabled(bool ignoreVolumeStack = false)
         {
-            return Application.isPlaying && camera.cameraType == CameraType.Game &&
-                   frameSettings.IsEnabled(FrameSettingsField.Volumetrics) &&
-                   frameSettings.IsEnabled(FrameSettingsField.ReprojectionForVolumetrics);
+            bool a = Fog.IsVolumetricFogEnabled(this, ignoreVolumeStack);
+            bool b = frameSettings.IsEnabled(FrameSettingsField.ReprojectionForVolumetrics);
+            bool c = camera.cameraType == CameraType.Game;
+            bool d = Application.isPlaying;
+
+            return a && b && c && d;
         }
 
         // Pass all the systems that may want to update per-camera data here.
@@ -276,6 +279,8 @@ namespace UnityEngine.Rendering.HighDefinition
         // Otherwise, previous frame view constants will be wrong.
         public void Update(FrameSettings currentFrameSettings, HDRenderPipeline hdrp, MSAASamples msaaSamples, XRPass xrPass)
         {
+            bool ignoreVolumeStack = true; // Unfortunately, it is initialized after this function call
+
             // store a shortcut on HDAdditionalCameraData (done here and not in the constructor as
             // we don't create HDCamera at every frame and user can change the HDAdditionalData later (Like when they create a new scene).
             camera.TryGetComponent<HDAdditionalCameraData>(out m_AdditionalCameraData);
@@ -287,9 +292,13 @@ namespace UnityEngine.Rendering.HighDefinition
 
             // Handle memory allocation.
             {
+                // Have to do this every frame in case the settings have changed.
+                // The condition inside controls whether we perform init/deinit or not.
+                hdrp.ReinitializeVolumetricBufferParams(this, ignoreVolumeStack);
+
                 bool isCurrentColorPyramidRequired = m_frameSettings.IsEnabled(FrameSettingsField.RoughRefraction) || m_frameSettings.IsEnabled(FrameSettingsField.Distortion);
                 bool isHistoryColorPyramidRequired = m_frameSettings.IsEnabled(FrameSettingsField.SSR) || antialiasing == AntialiasingMode.TemporalAntialiasing;
-                bool isVolumetricHistoryRequired   = IsVolumetricReprojectionEnabled();
+                bool isVolumetricHistoryRequired   = IsVolumetricReprojectionEnabled(ignoreVolumeStack);
 
                 int numColorPyramidBuffersRequired = 0;
                 if (isCurrentColorPyramidRequired)
@@ -300,11 +309,11 @@ namespace UnityEngine.Rendering.HighDefinition
                 int numVolumetricBuffersRequired = isVolumetricHistoryRequired ? 2 : 0; // History + feedback
 
                 if ((m_NumColorPyramidBuffersAllocated != numColorPyramidBuffersRequired) ||
-                    (m_NumVolumetricBuffersAllocated != numVolumetricBuffersRequired))
+                    (m_NumVolumetricBuffersAllocated   != numVolumetricBuffersRequired))
                 {
                     // Reinit the system.
                     colorPyramidHistoryIsValid = false;
-                    hdrp.DeinitializeVolumetricLightingPerCameraData(this);
+                    volumetricHistoryIsValid   = false;
 
                     // The history system only supports the "nuke all" option.
                     m_HistoryRTSystem.Dispose();
@@ -313,20 +322,16 @@ namespace UnityEngine.Rendering.HighDefinition
                     if (numColorPyramidBuffersRequired != 0)
                     {
                         AllocHistoryFrameRT((int)HDCameraFrameHistoryType.ColorBufferMipChain, HistoryBufferAllocatorFunction, numColorPyramidBuffersRequired);
-                        colorPyramidHistoryIsValid = false;
                     }
 
-                    hdrp.InitializeVolumetricLightingHistoryPerCamera(this, numVolumetricBuffersRequired);
+                    if (numVolumetricBuffersRequired != 0)
+                    {
+                        hdrp.AllocateVolumetricHistoryBuffers(this, numVolumetricBuffersRequired);
+                    }
 
                     // Mark as init.
                     m_NumColorPyramidBuffersAllocated = numColorPyramidBuffersRequired;
-                    m_NumVolumetricBuffersAllocated = numVolumetricBuffersRequired;
-                }
-
-                // Init the vbuffer params if were never initialized
-                if(vBufferParams == null)
-                {
-                    hdrp.InitializeVBufferParameters(this);
+                    m_NumVolumetricBuffersAllocated   = numVolumetricBuffersRequired;
                 }
             }
 
@@ -364,7 +369,7 @@ namespace UnityEngine.Rendering.HighDefinition
             UpdateAllViewConstants();
             isFirstFrame = false;
 
-            hdrp.UpdateVolumetricLightingPerCameraData(this);
+            hdrp.UpdateVolumetricBufferParams(this, ignoreVolumeStack);
 
             UpdateVolumeParameters();